dbeaver/dbeaver#8476 refactor open ai client and legacy api switch
