+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/copilot/CopilotCompletionEngine.java
@@ -18,18 +18,22 @@ package org.jkiss.dbeaver.model.ai.engine.copilot;
 
 import org.jkiss.code.NotNull;
 import org.jkiss.dbeaver.DBException;
-import org.jkiss.dbeaver.model.ai.*;
+import org.jkiss.dbeaver.model.ai.AIFunctionCall;
+import org.jkiss.dbeaver.model.ai.AIMessage;
+import org.jkiss.dbeaver.model.ai.AIMessageType;
+import org.jkiss.dbeaver.model.ai.AIUsage;
 import org.jkiss.dbeaver.model.ai.engine.*;
 import org.jkiss.dbeaver.model.ai.engine.copilot.dto.*;
-import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIClient;
 import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIConstants;
-import org.jkiss.dbeaver.model.ai.engine.openai.dto.*;
+import org.jkiss.dbeaver.model.ai.engine.openai.OpenAiUtils;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIMessage;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIResponsesResponse;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAITool;
 import org.jkiss.dbeaver.model.ai.internal.AIMessages;
 import org.jkiss.dbeaver.model.ai.utils.DisposableLazyValue;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.utils.CommonUtils;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
@@ -74,10 +78,11 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
         @NotNull DBRProgressMonitor monitor,
         @NotNull AIEngineRequest request
     ) throws DBException {
-        CopilotResponsesResponse chatResponse = client.getInstance().chat(
+        OAIResponsesResponse chatResponse = client.getInstance().chat(
             monitor,
             requestSessionToken(monitor).token(),
-            createOpenAiRequest(request)
+            createLegacyChatRequest(request, false),
+            OpenAiUtils.createOpenAiRequest(request, getModelName())
         );
 
         return toEngineResponse(chatResponse);
@@ -92,7 +97,8 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
         client.getInstance().createChatCompletionStream(
             monitor,
             requestSessionToken(monitor).token(),
-            createOpenAiRequest(request),
+            OpenAiUtils.createOpenAiRequest(request, getModelName()),
+            createLegacyChatRequest(request, true),
             listener
         );
     }
@@ -154,61 +160,9 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
             .build();
     }
 
-    @NotNull
-    private static List<OAIMessage> fromMessages(@NotNull List<AIMessage> messages) {
-        List<OAIMessage> result = new ArrayList<>(messages.size());
-        for (AIMessage message : messages) {
-            if (message.getFunctionCall() != null) {
-                OAIMessage functionCallMessage = OAIMessageFactory.fromAIMessage(message);
-                if (!CommonUtils.isEmpty(functionCallMessage.callId)) {
-                    result.add(functionCallMessage);
-                    // OpenAI Responses API requires matching function_call_output for each function_call.
-                    // Keep orphan function calls in history as regular assistant messages.
-                    result.add(OAIMessageFactory.fromAIMessage(message, functionCallMessage.callId));
-                }
-            } else {
-                result.add(OAIMessageFactory.fromAIMessage(message));
-            }
-        }
-        return result;
-    }
 
     @NotNull
-    private CopilotResponsesRequest createOpenAiRequest(@NotNull AIEngineRequest request) throws DBException {
-        CopilotResponsesRequest oaiRequest = new CopilotResponsesRequest();
-        List<AIMessage> messages = request.getMessages();
-        oaiRequest.input = fromMessages(messages);
-        oaiRequest.store = false;
-        oaiRequest.model = getModelName();
-
-        if (!CommonUtils.isEmpty(request.getFunctions())) {
-            List<OAITool> tools = new ArrayList<>();
-            for (AIFunctionDescriptor fd : request.getFunctions()) {
-                OAITool tool = new OAITool();
-                tool.type = OAITool.TYPE_FUNCTION;
-                tool.name = fd.getFullId();
-                tool.description = fd.getAiDescription();
-                tool.parameters.type = OAIToolParameters.TYPE_OBJECT;
-                List<String> requiredFields = new ArrayList<>();
-                for (AIFunctionParameter param : fd.getParameters()) {
-                    OAIToolParameter tp = new OAIToolParameter();
-                    tp.type = param.getType();
-                    tp.description = param.getDescription();
-                    tp.enumItems = param.getValidValues();
-                    requiredFields.add(param.getName());
-                    tool.parameters.properties.put(param.getName(), tp);
-                }
-                tool.parameters.required = requiredFields.toArray(new String[0]);
-                tools.add(tool);
-            }
-            oaiRequest.tools = tools;
-        }
-
-        return oaiRequest;
-    }
-
-    @NotNull
-    private AIEngineResponse toEngineResponse(@NotNull CopilotResponsesResponse response) throws DBException {
+    private AIEngineResponse toEngineResponse(@NotNull OAIResponsesResponse response) throws DBException {
         List<OAIMessage> messages = response.output.stream()
             .filter(msg -> !OAIMessage.TYPE_FUNCTION_REASONING.equals(msg.type))
             .toList();
@@ -222,7 +176,7 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
         }
         OAIMessage message = messages.getFirst();
         if (OAIMessage.TYPE_FUNCTION_CALL.equals(message.type)) {
-            AIFunctionCall fc = OpenAIClient.createFunctionCall(message);
+            AIFunctionCall fc = OpenAiUtils.createFunctionCall(message);
             return new AIEngineResponse(fc, usage);
         } else {
             List<String> choices = messages.stream()
