--- plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/copilot/dto/CopilotResponsesResponse.java
@@ -1,29 +1,0 @@
-/*
- * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2026 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jkiss.dbeaver.model.ai.engine.copilot.dto;
-
-import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIMessage;
-import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIResponsesBase;
-
-import java.util.List;
-
-public class CopilotResponsesResponse extends OAIResponsesBase {
-    public List<OAIMessage> output;
-
-    public CopilotResponsesResponse() {
-    }
-}
