+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/openai/OpenAIClientLegacy.java
@@ -24,6 +24,7 @@ import org.jkiss.dbeaver.model.ai.AIMessage;
 import org.jkiss.dbeaver.model.ai.AIMessageMeta;
 import org.jkiss.dbeaver.model.ai.AIMessageType;
 import org.jkiss.dbeaver.model.ai.AIMetaTypes;
+import org.jkiss.dbeaver.model.ai.engine.AIEngineResponseConsumer;
 import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIMessageFactory;
 import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIResponsesRequest;
 import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIResponsesResponse;
@@ -38,8 +39,9 @@ import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
 import java.util.Locale;
+import java.util.function.Consumer;
 
-public class OpenAIClientLegacy extends OpenAIClient {
+public class OpenAIClientLegacy extends OpenAiClientBase {
     private static final Duration TIMEOUT = Duration.ofSeconds(30);
     private static final Gson GSON = new GsonBuilder().create();
 
@@ -110,5 +112,24 @@ public class OpenAIClientLegacy extends OpenAIClient {
                 )
             )).toList();
         return oaiResponse;
+    }
+
+    @Override
+    public void createChatCompletionStream(
+        @NotNull DBRProgressMonitor monitor,
+        @NotNull OAIResponsesRequest completionRequest,
+        @NotNull AIEngineResponseConsumer listener
+    ) throws DBException {
+        HttpRequest request = createCompletionRequest(completionRequest);
+
+        HttpRequest modifiedRequest = applyFilters(request);
+
+        Consumer<String> stringConsumer = new OpenAiAPIStreamConsumer(listener);
+        client.sendAsync(
+            modifiedRequest,
+            stringConsumer,
+            listener::error,
+            listener::completeBlock
+        );
     }
 }
