+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/openai/OpenAIRequestFilter.java
@@ -21,7 +21,7 @@ import org.jkiss.utils.HttpConstants;
 
 import java.net.http.HttpRequest;
 
-public class OpenAIRequestFilter implements OpenAIClient.HttpRequestFilter {
+public class OpenAIRequestFilter implements OpenAiClientBase.HttpRequestFilter {
     private final String token;
 
     public OpenAIRequestFilter(String token) {
