+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/openai/OpenAiClientBase.java
@@ -1,0 +1,129 @@
+/*
+ * DBeaver - Universal Database Manager
+ * Copyright (C) 2010-2026 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jkiss.dbeaver.model.ai.engine.openai;
+
+import com.google.gson.Gson;
+import org.jkiss.code.NotNull;
+import org.jkiss.code.Nullable;
+import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.Log;
+import org.jkiss.dbeaver.model.ai.engine.AIEngineResponseConsumer;
+import org.jkiss.dbeaver.model.ai.engine.AbstractHttpAIClient;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIModel;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIModelList;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIResponsesRequest;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIResponsesResponse;
+import org.jkiss.dbeaver.model.ai.utils.AIHttpUtils;
+import org.jkiss.dbeaver.model.data.json.JSONUtils;
+import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
+import org.jkiss.dbeaver.utils.GeneralUtils;
+import org.jkiss.utils.HttpConstants;
+
+import java.net.http.HttpRequest;
+import java.time.Duration;
+import java.util.List;
+
+public abstract class OpenAiClientBase extends AbstractHttpAIClient {
+    protected static final Duration TIMEOUT = Duration.ofSeconds(30);
+    protected static final Gson GSON = JSONUtils.GSON;
+    private static final Log log = Log.getLog(OpenAiClientBase.class);
+    protected final String baseUrl;
+    protected final List<OpenAIClient.HttpRequestFilter> requestFilters;
+
+    public OpenAiClientBase(
+        @NotNull String baseUrl,
+        @NotNull List<OpenAIClient.HttpRequestFilter> requestFilters
+    ) {
+        if (!baseUrl.endsWith("/")) {
+            baseUrl += "/";
+        }
+        this.baseUrl = baseUrl;
+        this.requestFilters = requestFilters;
+    }
+
+    @NotNull
+    protected static String serializeValue(@Nullable Object value) throws DBException {
+        try {
+            return GSON.toJson(value);
+        } catch (Exception e) {
+            throw new DBException("Error serializing value", e);
+        }
+    }
+
+    @NotNull
+    public abstract OAIResponsesResponse createChatCompletion(
+        @NotNull DBRProgressMonitor monitor,
+        @NotNull OAIResponsesRequest completionRequest
+    ) throws DBException;
+
+    public abstract void createChatCompletionStream(
+        @NotNull DBRProgressMonitor monitor,
+        @NotNull OAIResponsesRequest completionRequest,
+        @NotNull AIEngineResponseConsumer listener
+    ) throws DBException;
+
+    @NotNull
+    public List<OAIModel> getModels(@NotNull DBRProgressMonitor monitor) throws DBException {
+        HttpRequest request = HttpRequest.newBuilder()
+            .uri(AIHttpUtils.resolve(baseUrl, "models"))
+            .GET()
+            .timeout(TIMEOUT)
+            .build();
+
+        HttpRequest modifiedRequest = applyFilters(request);
+        return GSON.fromJson(client.send(monitor, modifiedRequest), OAIModelList.class).data();
+    }
+
+    @NotNull
+    public HttpRequest applyFilters(@NotNull HttpRequest request) throws DBException {
+        return applyFilters(request, true);
+    }
+
+    @NotNull
+    public HttpRequest applyFilters(@NotNull HttpRequest request, boolean setContentType) throws DBException {
+        for (HttpRequestFilter filter : requestFilters) {
+            request = filter.filter(request, setContentType);
+        }
+        return request;
+    }
+
+    @NotNull
+    @Override
+    protected DBException mapHttpError(int statusCode, @NotNull String body) {
+        log.debug("OpenAI request failed: " + statusCode + ", " + body);
+        return new DBException("OpenAI request failed: " + AIHttpUtils.parseOpenAIStyleErrorMessage(body));
+    }
+
+    @NotNull
+    protected HttpRequest createCompletionRequest(@NotNull OAIResponsesRequest completionRequest) throws DBException {
+        return HttpRequest.newBuilder()
+            .uri(AIHttpUtils.resolve(baseUrl, OpenAIConstants.ENDPOINT_RESPONSES))
+            .header(HttpConstants.HEADER_USER_AGENT, GeneralUtils.getProductTitle())
+            .POST(HttpRequest.BodyPublishers.ofString(serializeValue(completionRequest)))
+            .timeout(TIMEOUT)
+            .build();
+    }
+
+    /**
+     * Interface for filtering and modifying HTTP requests before they are sent.
+     * Implementations can be used to add authentication headers, modify the request body, etc.
+     */
+    public interface HttpRequestFilter {
+        @NotNull
+        HttpRequest filter(@NotNull HttpRequest request, boolean setContentType) throws DBException;
+    }
+}
